backend: Remove obsolete get_type() function of adapters
The get_type() function of the adapters is not needed and only contains
duplicated information, as the mapping is defined in the __init__.py files
and the type of get_type() is only used for the API output. But the type
exists also in the configuration and can be used from there.